LayoutTests/fast/custom-elements/disconnected-callbacks.html

<!DOCTYPE html>
<html>
<head>
<title>Custom Elements: disconnectedCallback</title>
<meta name="author" title="Ryosuke Niwa" href="mailto:rniwa@webkit.org">
<meta name="assert" content="disconnectedCallback must be enqueued whenever custom element is removed from a document">
<link rel="help" href="https://w3c.github.io/webcomponents/spec/custom/#dfn-connected-callback">
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<link rel='stylesheet' href='../../resources/testharness.css'>
</head>
<body>
<div id="log"></div>
<script>

var calls = [];
class MyCustomElement extends HTMLElement {
    connectedCallback() { calls.push('connected', this); }
    disconnectedCallback() { calls.push('disconnected', this); }
}
customElements.define('my-custom-element', MyCustomElement);

test(function () {
    var instance = document.createElement('my-custom-element');
    document.body.appendChild(instance);

    calls = [];
    document.body.removeChild(instance);
    assert_array_equals(calls, ['disconnected', instance]);
}, 'Removing a custom element from a document must enqueue and invoke disconnectedCallback');

test(function () {
    var instance = document.createElement('my-custom-element');
    var parent = document.createElement('div');
    parent.appendChild(instance);

    calls = [];
    parent.removeChild(instance);
    assert_array_equals(calls, []);
}, 'Removing a custom element from a detahed node must not enqueue and invoke connectedCallback');

test(function () {
    var instance = document.createElement('my-custom-element');
    var documentWithoutWindow = document.implementation.createHTMLDocument();
    documentWithoutWindow.body.appendChild(instance);

    calls = [];
    documentWithoutWindow.body.removeChild(instance);
    assert_array_equals(calls, ['disconnected', instance]);
}, 'Removing a custom element from a window-less document must enqueue and invoke disconnectedCallback');

test(function () {
    var instance = document.createElement('my-custom-element');
    var parent = document.createElement('div');
    parent.appendChild(instance);
    document.body.appendChild(parent);

    calls = [];
    document.body.removeChild(parent);
    assert_array_equals(calls, ['disconnected', instance]);
}, 'Removing an ancestor of a custom element from a document must enqueue and invoke disconnectedCallback');

test(function () {
    var instance = document.createElement('my-custom-element');
    var parent = document.createElement('div');
    parent.appendChild(instance);
    var documentWithoutWindow = document.implementation.createHTMLDocument();
    documentWithoutWindow.body.appendChild(parent);

    calls = [];
    documentWithoutWindow.body.removeChild(parent);
    assert_array_equals(calls, ['disconnected', instance]);
}, 'Removing an ancestor of custom element from a a window-less document must enqueue and invoke disconnectedCallback');

test(function () {
    var instance = document.createElement('my-custom-element');
    var host = document.createElement('div');
    document.body.appendChild(host);
    var shadowRoot = host.attachShadow({mode: 'closed'});
    shadowRoot.appendChild(instance);

    calls = [];
    shadowRoot.removeChild(instance);
    assert_array_equals(calls, ['disconnected', instance]);
}, 'Removing a custom element from a connected shadow tree must enqueue and invoke disconnectedCallback');

test(function () {
    var instance = document.createElement('my-custom-element');
    var host = document.createElement('div');
    var shadowRoot = host.attachShadow({mode: 'closed'});
    shadowRoot.appendChild(instance);
    document.body.appendChild(host);

    calls = [];
    document.body.removeChild(host);
    assert_array_equals(calls, ['disconnected', instance]);
}, 'Removing the shadow host of a shadow tree with a custom element from a document must enqueue and invoke disconnectedCallback');

test(function () {
    var instance = document.createElement('my-custom-element');
    var host = document.createElement('div');
    var shadowRoot = host.attachShadow({mode: 'closed'});
    shadowRoot.appendChild(instance);

    calls = [];
    shadowRoot.removeChild(instance);
    assert_array_equals(calls, []);
}, 'Removing a custom element from a detached shadow tree must not enqueue and invoke disconnectedCallback');


</script>
</body>
</html>
